- Add route for upload avatars - Add static folder - Routes refactor
