+++ src/main/kotlin/com/example/data/source/AuthDataSource.kt
@@ -1,0 +1,16 @@
+package com.example.data.source
+
+import com.example.data.model.user.Friend
+import com.example.data.model.user.NewUserInfo
+import com.example.data.model.user.UserDTO
+import com.example.data.model.user.UserFromId
+
+interface AuthDataSource {
+
+    suspend fun regUser(userCredentials: UserDTO): String?
+
+    suspend fun login(userCredentials: UserDTO): String?
+
+    suspend fun logout(uid: String): Boolean
+
+}
