+++ src/main/kotlin/com/example/data/source/UserDataSource.kt
@@ -2,17 +2,9 @@ package com.example.data.source
 
 import com.example.data.model.user.Friend
 import com.example.data.model.user.NewUserInfo
-import com.example.data.model.user.UserDTO
 import com.example.data.model.user.UserFromId
 
 interface UserDataSource {
-
-    suspend fun regUser(userCredentials: UserDTO): String?
-
-    suspend fun login(userCredentials: UserDTO): String?
-
-    suspend fun logout(uid: String): Boolean
-
     suspend fun getUserById(uid: String): UserFromId?
 
     suspend fun friendshipRequest(selfId: String, userId: String): Boolean
