+++ src/main/kotlin/com/example/data/source/UserDataSourceImpl.kt
@@ -1,59 +1,18 @@
 package com.example.data.source
 
-import com.example.data.model.user.*
-import com.example.utils.Constants.FOLLOWERS
-import com.example.utils.Constants.FRIENDS
-import com.example.utils.Constants.FRIENDSHIPS_REQUESTS
+import com.example.data.model.user.Friend
+import com.example.data.model.user.NewUserInfo
+import com.example.data.model.user.User
+import com.example.data.model.user.UserFromId
+import com.example.utils.Constants
 import org.litote.kmongo.coroutine.CoroutineDatabase
 import org.litote.kmongo.eq
-import org.mindrot.jbcrypt.BCrypt
 
 class UserDataSourceImpl(
     private val db: CoroutineDatabase,
 ) : UserDataSource {
 
     private val users = db.getCollection<User>()
-
-    override suspend fun regUser(userCredentials: UserDTO): String? {
-        return if (users.find(User::username eq userCredentials.username).toList().isEmpty()) {
-            users.insertOne(
-                User(
-                    username = userCredentials.username,
-                    password = userCredentials.hashedPassword(),
-                    selfInfo = "",
-                    onlineStatus = true,
-                    lastActionTime = System.currentTimeMillis(),
-                    timestamp = System.currentTimeMillis(),
-                    friends = mutableListOf(),
-                    followers = mutableListOf(),
-                    friendshipRequests = mutableListOf(),
-                )
-            )
-            users.find(User::username eq userCredentials.username).first()?.id
-        } else null
-    }
-
-    override suspend fun login(userCredentials: UserDTO): String? {
-        val user = users.find(User::username eq userCredentials.username).first()
-        return if (user != null) {
-            if (BCrypt.checkpw(userCredentials.password, user.password)) {
-                user.onlineStatus = true
-                user.lastActionTime = System.currentTimeMillis()
-                users.updateOne(User::username eq user.username, user)
-                user.id
-            } else ""
-        } else ""
-    }
-
-    override suspend fun logout(uid: String): Boolean {
-        val user = users.find(User::id eq uid).first()
-        user?.let {
-            it.onlineStatus = false
-            it.lastActionTime = System.currentTimeMillis()
-            users.updateOne(User::id eq uid, it)
-            return true
-        } ?: return false
-    }
 
     override suspend fun getUserById(uid: String): UserFromId? {
         users.find(
@@ -114,19 +73,19 @@ class UserDataSourceImpl(
         val list = mutableListOf<Friend>()
         users.find(User::id eq uid).first()?.let {
             when (type) {
-                FRIENDS -> {
+                Constants.FRIENDS -> {
                     it.friends.forEach { uid ->
                         list.add(getFriend(uid)!!)
                     }
                 }
 
-                FOLLOWERS -> {
+                Constants.FOLLOWERS -> {
                     it.followers.forEach { uid ->
                         list.add(getFriend(uid)!!)
                     }
                 }
 
-                FRIENDSHIPS_REQUESTS -> {
+                Constants.FRIENDSHIPS_REQUESTS -> {
                     it.friendshipRequests.forEach { uid ->
                         list.add(getFriend(uid)!!)
                     }
