+++ src/main/kotlin/com/example/di/MainModule.kt
@@ -1,9 +1,10 @@
 package com.example.di
 
 import com.example.data.chat.ChatController
-import com.example.data.room.RoomController
+import com.example.data.controllers.RoomController
 import com.example.data.source.*
-import com.example.data.users.UserController
+import com.example.data.controllers.AuthController
+import com.example.data.controllers.UserController
 import org.koin.dsl.module
 import org.litote.kmongo.coroutine.coroutine
 import org.litote.kmongo.reactivestreams.KMongo
@@ -14,8 +15,8 @@ val mainModule = module {
             .coroutine
             .getDatabase("chat_db")
     }
-    single<UserDataSource> {
-        UserDataSourceImpl(get())
+    single<AuthDataSource> {
+        AuthDataSourceImpl(get())
     }
     single<MessageDataSource> {
         MessageDataSourceImpl(get())
@@ -23,13 +24,19 @@ val mainModule = module {
     single<ChatDataSource> {
         ChatDataSourceImpl(get())
     }
+    single<UserDataSource> {
+        UserDataSourceImpl(get())
+    }
     single {
-        UserController(get())
+        AuthController(get())
     }
     single {
         RoomController(get())
     }
     single {
         ChatController(get())
+    }
+    single {
+        UserController(get())
     }
 }
